+++ src/meeple/util/api_util.py
@@ -1,3 +1,4 @@
+import json
 import time
 from typing import List
 
@@ -10,24 +11,47 @@ from meeple.util.message_util import error_msg
 from meeple.util.output_util import CustomProgress
 
 BGG_DOMAIN = "boardgamegeek.com"
-API2_BASE_URL = f"https://{BGG_DOMAIN}/xmlapi2"
+
+JSON_API_BASE_URL = "https://api.geekdo.com/api"
+XML2_API_BASE_URL = f"https://{BGG_DOMAIN}/xmlapi2"
 
 # BoardGameGeek item types
 BOARDGAME_TYPE = "boardgame"
 EXPANSION_TYPE = "boardgameexpansion"
 
 
-def _bgg_api_get(endpoint: str, params: dict) -> List[Item]:
+def _bgg_json_api_get(endpoint: str, params: dict = {}) -> dict:
+    """Get items from provided GeekDo JSON API endpoint.
+
+    Args:
+        endpoint (str): GEEKDO API endpoint.
+        params (dict): Request parameters.
+
+    Returns:
+        response_dict: Dictionary representation of response.
+    """
+    response = requests.get(f"{JSON_API_BASE_URL}/{endpoint}", params=params)
+    match response.status_code:
+        case 200:
+            return json.loads(response.content)
+        case _:
+            error_msg(
+                f"Unknown GeekDo API Error | HTTP {response.status_code} | {response.content}"
+            )
+
+
+def _bgg_xml2_api_get(endpoint: str, params: dict = {}) -> List[Item]:
     """Get items from the provided BGG endpoint.
 
     Args:
         endpoint (str): BGG endpoint.
+        params (dict): Request parameters.
 
     Returns:
-        List[Item]: List of items.
+        response_dict: Dictionary representation of response.
     """
     while True:
-        response = requests.get(f"{API2_BASE_URL}/{endpoint}", params=params)
+        response = requests.get(f"{XML2_API_BASE_URL}/{endpoint}", params=params)
         match response.status_code:
             case 200:
                 # successful return
@@ -44,7 +68,7 @@ def _bgg_api_get(endpoint: str, params: dict) -> List[Item]:
                 # unknown api error
                 # TODO: log this error and print out a friendlier message to the user
                 error_msg(
-                    f"Unknown API Error | HTTP {response.status_code} | {response.content}"
+                    f"Unknown BGG API Error | HTTP {response.status_code} | {response.content}"
                 )
 
 
@@ -69,7 +93,7 @@ def get_bgg_items(bgg_ids: List[int]) -> List[Item]:
         "type": f"{BOARDGAME_TYPE},{EXPANSION_TYPE}",
         "stats": 1,
     }
-    resp_dict = _bgg_api_get(endpoint="thing", params=params)
+    resp_dict = _bgg_xml2_api_get(endpoint="thing", params=params)
     return _serialize_resp_dict(resp_dict, Item)
 
 
@@ -95,7 +119,7 @@ def get_bgg_hot() -> List[Item]:
         List[Item]: List of items.
     """
     params = {"type": BOARDGAME_TYPE}
-    resp_dict = _bgg_api_get(endpoint="hot", params=params)
+    resp_dict = _bgg_xml2_api_get(endpoint="hot", params=params)
     return _serialize_resp_dict(resp_dict, Item)
 
 
@@ -109,7 +133,7 @@ def search_bgg(query: str) -> List[Item]:
         List[Item]: List of items.
     """
     params = {"type": BOARDGAME_TYPE, "query": query}
-    resp_dict = _bgg_api_get(endpoint="search", params=params)
+    resp_dict = _bgg_xml2_api_get(endpoint="search", params=params)
     return _serialize_resp_dict(resp_dict, Item)
 
 
@@ -123,11 +147,11 @@ def get_bgg_user(username: str) -> BGGUser:
         BGGUser: BGGUser.
     """
     params = {"name": username}
-    resp_dict = _bgg_api_get(endpoint="user", params=params)
+    resp_dict = _bgg_xml2_api_get(endpoint="user", params=params)
     return BGGUser.from_bgg_dict(resp_dict["user"])
 
 
-def get_bgg_user_collection(username: str):
+def get_bgg_user_collection(username: str) -> List[BGGCollectionItem]:
     """Get BGG User Collection by username.
 
     Args:
@@ -139,5 +163,9 @@ def get_bgg_user_collection(username: str):
     params = {"username": username, "brief": 1}
     with CustomProgress() as progress:
         progress.add_task("Fetching collection", start=False, total=None)
-        resp_dict = _bgg_api_get(endpoint="collection", params=params)
+        resp_dict = _bgg_xml2_api_get(endpoint="collection", params=params)
     return _serialize_resp_dict(resp_dict, BGGCollectionItem)
+
+
+def get_campaigns() -> dict:
+    return _bgg_json_api_get(endpoint="ending_preorders")